+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- When creating a Connect credential using token authentication or an API key
+  the server URL is automatically discovered even when extra paths are provided
+  (e.g., `/connect`). (#2999)
+
 ## [1.24.0]
 
 ### Added